feat: implementa CRUD de Campeonato

+++ backend/src/main/java/com/champsched/repository/CampeonatoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CampeonatoRepository extends JpaRepository<Campeonato, Long> {
+public interface CampeonatoRepository extends JpaRepository<Campeonato, Integer> {
     
 }
